Compiling indexing assignment
CCompiler2: Add logic to compile indexing assignment
Typechecker: Update AST model to accurately represent the types of
index-expressions that can be used in an indexing assignment (ie. range
index-expressions are forbidden).

+++ abra_core/src/transpile/targetv2/c/abra/include/prelude.h
@@ -96,7 +96,7 @@ AbraBool* AbraString__eq(size_t nargs, AbraString* self, AbraAny* other);
 AbraString* AbraString__concat(size_t nargs, AbraString* self, AbraAny* other);
 
 AbraArray* AbraArray_make_with_capacity(size_t length, size_t cap);
-AbraUnit AbraArray_set(AbraArray* self, size_t index, AbraAny* item);
+AbraUnit AbraArray_set(AbraArray* self, int64_t index, AbraAny* item);
 AbraAny* AbraArray_get(AbraArray* self, int64_t index);
 AbraArray* AbraArray_slice(AbraArray* self, int64_t index);
 AbraArray* AbraArray_get_range(AbraArray* self, int64_t start, int64_t end);
